fix: temporary fix for new nodejs
This is annoying. Newer Node.js is getting picky about using require on resources inside a type=module. The best solution would be to migrate everything to ESM but I don't have a testing solution that is all ESM yet so, for now, backing out type=module and using .mjs