feat: working runner/loader
This is a bit of a hack. There's a bug somewhere in the Node.js loader infra that crashes when you try to do IO inside a loader script. This does a full stage/deflate before running in order to have the entire dep tree available without doing additional loader IO.